update error solution

+++ app/Http/Controllers/Api/UserController.php
@@ -28,8 +28,10 @@ class UserController extends Controller
         $url = 'https://s3.' . env('AWS_DEFAULT_REGION') . '.amazonaws.com/' . env('AWS_BUCKET') . '/';
         $userImage = $user->userImage;
         $vehicle = $user->vehicle;
+        $vehicle->type = json_decode($vehicle->type );
         $transaction = $user->transactions;
         $trip = $user->trip;
+
         $notification = $user->notifications;
 
         return response()->json(['user' => $user,'url'=>$url, 'success' => true], 200);
@@ -37,6 +39,9 @@ class UserController extends Controller
 
     public function update(Request $request)
     {
+
+        // dd($request);
+
 
         $validator = Validator::make($request->all(), [
             'name' => 'nullable|string|max:255',
@@ -78,12 +83,15 @@ class UserController extends Controller
         // update profile image
         if ($request->hasFile('profile_image')) {
             
-            if (Storage::disk('s3')->exists($userImage->profile_image)) {
-           
+            $userImage = $user->userImage;  // Assuming the User model has a relationship to UserImage
+        
+            // Check if the user's current profile image exists on S3 and delete it
+            if ($userImage && !is_null($userImage->profile_image) && Storage::disk('s3')->exists($userImage->profile_image)) {
                 Storage::disk('s3')->delete($userImage->profile_image);
             }
+            
             $profileImage = $request->file('profile_image');
-            $profileImageName = time() . '_' . $user->nrc_no . '.' . $profileImage->getClientOriginalExtension();
+            $profileImageName = uniqid() . '_' . $user->nrc_no . '.' . $profileImage->getClientOriginalExtension();
             Storage::disk('s3')->put($profileImageName, file_get_contents($profileImage));
             $userImage->profile_image = $profileImageName;
             $userImage->save();
@@ -97,7 +105,7 @@ class UserController extends Controller
                 Storage::disk('s3')->delete($userImage->front_nrc_image);
             }
             $frontNrcImage = $request->file('front_nrc_image');
-            $frontNrcImageName = time() . '.' . $frontNrcImage->getClientOriginalExtension();
+            $frontNrcImageName = uniqid() . '.' . $frontNrcImage->getClientOriginalExtension();
             Storage::disk('s3')->put($frontNrcImageName, file_get_contents($frontNrcImage));
             $userImage->front_nrc_image = $frontNrcImageName;
             $userImage->save();
@@ -111,7 +119,7 @@ class UserController extends Controller
                 Storage::disk('s3')->delete($userImage->back_nrc_image);
             }
             $backNrcImage = $request->file('back_nrc_image');
-            $backNrcImageName = time() . '.' . $backNrcImage->getClientOriginalExtension();
+            $backNrcImageName = uniqid() . '.' . $backNrcImage->getClientOriginalExtension();
             Storage::disk('s3')->put($backNrcImageName, file_get_contents($backNrcImage));
             $userImage->back_nrc_image = $backNrcImageName;
             $userImage->save();
@@ -125,7 +133,7 @@ class UserController extends Controller
                 Storage::disk('s3')->delete($userImage->front_license_image);
             }
             $backNrcImage = $request->file('front_license_image');
-            $backNrcImageName = time() . '.' . $backNrcImage->getClientOriginalExtension();
+            $backNrcImageName = uniqid() . '.' . $backNrcImage->getClientOriginalExtension();
             Storage::disk('s3')->put($backNrcImageName, file_get_contents($backNrcImage));
             $userImage->front_license_image = $backNrcImageName;
             $userImage->save();
@@ -138,7 +146,7 @@ class UserController extends Controller
                 Storage::disk('s3')->delete($userImage->back_license_image);
             }
             $backLicenseImage = $request->file('back_license_image');
-            $backLicenseImageName = time() . '.' . $backLicenseImage->getClientOriginalExtension();
+            $backLicenseImageName = uniqid() . '.' . $backLicenseImage->getClientOriginalExtension();
             Storage::disk('s3')->put($backLicenseImageName, file_get_contents($backLicenseImage));
           
             $userImage->back_license_image = $backLicenseImageName;
@@ -157,6 +165,7 @@ class UserController extends Controller
 
         if ($user->vehicle) {
             $vehicle = Vehicle::where('user_id', $user->id)->first();
+           
         } else {
             $vehicle = new Vehicle();
             $vehicle->user_id = $user->id;
@@ -183,7 +192,10 @@ class UserController extends Controller
 
             $vehicle->vehicle_image_url = $vehicleImageName;
         }
+        
         $vehicle->save();
+        $vehicle->type = json_decode($vehicle->type);
+        
         return response()->json(['user' => $user, 'status' => 'User updated successfully', 'success' => true], 200);
     }
 
@@ -296,6 +308,8 @@ class UserController extends Controller
                     'start_time'=>$trip->start_time,
                     'end_time' => $trip->end_time,
                     'extra_fee_list'=>$fees->toArray(),
+                    'polyline' => json_decode($trip->polyline),
+                    'commission_fee'=>$trip->commission_fee,
                     // 'extra_fee_list'=>$extraFees,
 
                     'created_at' => Carbon::parse($trip->created_at)->format('Y-m-d h:i A'),
@@ -372,6 +386,255 @@ class UserController extends Controller
         return response()->json(['user' => $user, 'status' => 'User updated successfully', 'success' => true], 200);
     }
 
+    public function totalTripandPrice($range){
+        $user = Auth::user();
+        $trips = $user->trips;
 
+        
+
+        if ($range === 'day') {
+            $startDate = Carbon::now()->startOfDay();
+            $endDate = Carbon::now()->endOfDay();
+
+            // Get trips for the current day
+            $todayTrips = $trips->whereBetween('created_at', [$startDate, $endDate]);
+
+            $totalTripCount = $todayTrips->count();
+            $totalAmount = $todayTrips->sum('total_cost');
+
+            return response()->json([
+                'trip_count' => $totalTripCount,
+                'total_amount' => $totalAmount,
+            ]);
+
+            
+        } elseif ($range === 'week') {
+            $startDate = Carbon::now()->startOfWeek();
+            $endDate = Carbon::now()->endOfWeek();
+
+            // Get trips for the current day
+            $todayTrips = $trips->whereBetween('created_at', [$startDate, $endDate]);
+
+            $totalTripCount = $todayTrips->count();
+            $totalAmount = $todayTrips->sum('total_cost');
+
+            return response()->json([
+                'trip_count' => $totalTripCount,
+                'total_amount' => $totalAmount,
+            ]);
+        } elseif ($range === 'month') {
+            $startDate = Carbon::now()->startOfMonth();
+            $endDate = Carbon::now()->endOfMonth();
+
+            
+            // Get trips for the current day
+            $todayTrips = $trips->whereBetween('created_at', [$startDate, $endDate]);
+
+            $totalTripCount = $todayTrips->count();
+            $totalAmount = $todayTrips->sum('total_cost');
+
+            return response()->json([
+                'trip_count' => $totalTripCount,
+                'total_amount' => $totalAmount,
+            ]);
+
+        } elseif ($range === 'year') {
+            $startDate = Carbon::now()->startOfYear();
+            $endDate = Carbon::now()->endOfYear();
+
+            
+            // Get trips for the current day
+            $todayTrips = $trips->whereBetween('created_at', [$startDate, $endDate]);
+
+            $totalTripCount = $todayTrips->count();
+            $totalAmount = $todayTrips->sum('total_cost');
+
+            return response()->json([
+                'trip_count' => $totalTripCount,
+                'total_amount' => $totalAmount,
+            ]);
+        }
+
+        $tripCount = $trips->count();
+        $totalBalance = $trips->sum('total_cost');
+
+        return response()->json([
+            'trip_count' => $tripCount,
+            'total_balance' => $totalBalance,
+        ]);
+
+
+
+    }
+
+    public function topLists(){
+        // $authUser = Auth::user();
+
+        // // Drivers and their trip counts
+        // $drivers = Trip::where('status','completed')->select('driver_id', DB::raw('count(*) as total_trips'))
+        //                     ->groupBy('driver_id')
+        //                     ->orderBy('total_trips', 'desc')
+        //                     ->get();
+
+        // // Add ranking to drivers
+        // $driverDetails = $drivers->map(function ($driver, $index) {
+        //     $user = User::find($driver->driver_id);
+
+        //     return [
+        //         'driver_id' => $user->driver_id,
+        //         'name' => $user->name,
+        //         'email' => $user->email,
+        //         'total_trips' => $driver->total_trips,
+        //         'profile_image' =>  optional($user->userImage)->profile_image,
+        //         'rank' => $index + 1 // Add rank field based on index
+        //     ];
+        // });
+
+        // // Auth User's Rank
+        // $authUserRank = $driverDetails->firstWhere('driver_id', $authUser->driver_id)['rank'] ?? null;
+
+        // Top 10 Drivers
+        // $topDrivers = $driverDetails->take(10);
+
+        // Response ပြန်ပေးရန်
+        // return response()->json([
+        //     'auth_user' => [
+        //         'driver_id' => $authUser->driver_id,
+        //         'name' => $authUser->name,
+        //         'total_trips' => $authUser->trips->where('status','completed')->count(),
+        //         'profile_image' =>optional($authUser->userImage)->profile_image,
+        //         'rank' => $authUserRank,
+                
+        //     ],
+        //     'top_drivers' => $topDrivers,
+        //     'url'=>'https://s3.' . env('AWS_DEFAULT_REGION') . '.amazonaws.com/' . env('AWS_BUCKET') . '/',
+        // ]);
+
+        
+        // $startDate = Carbon::now()->startOfYear();
+        // $endDate = Carbon::now()->endOfYear();
+        
+        // // Fetch trips with completed status within the date range
+        // $trips = Trip::
+        //     whereBetween('created_at', [$startDate, $endDate])
+        //     ->orderBy('created_at')
+        //     ->get();
+        
+        // $collection = collect($trips);
+        
+        // // Group trips by month
+        // $trips = $collection->groupBy(function ($trip) {
+        //     return Carbon::parse($trip->created_at)->format('Y-m');
+        // })->map(function ($group) {
+        //     // Group by driver_id and get drivers with total trips
+        //     $drivers = $group->groupBy('driver_id')->map(function ($driverTrips, $driverId) {
+        //         $user = User::find($driverId);
+                
+        //         // Check if the user exists and has the 'user' role
+        //         if ($user && $user->roles()->where('name', 'user')->exists()) {
+        //             return [
+        //                 'driver_id' => $user->driver_id,
+        //                 'name' => $user->name,
+                        
+        //                 'total_trips' => $driverTrips->count(),
+        //                 'profile_image' => optional($user->userImage)->profile_image,
+        //             ];
+        //         }
+        //         return null;
+        //     })->filter()->sortByDesc('total_trips')->values(); // Sort and re-index
+        
+        //     // Add ranking to drivers
+        //     $rankedDrivers = $drivers->map(function ($driver, $index) {
+        //         $driver['rank'] = $index + 1; // Add rank based on index
+        //         return $driver;
+        //     }); // Take top 10
+
+        //     $authUser = Auth::user();
+        //     $authUserRank = $rankedDrivers->firstWhere('driver_id', $authUser->driver_id);
+
+        //     $auth = [
+               
+        //         'driver_id' => $authUser->driver_id,
+        //         'name' => $authUser->name,
+        //         'total_trips' => $authUser->trips->where('status','completed')->count(),
+        //         'profile_image' =>optional($authUser->userImage)->profile_image,
+        //         'auth_user_rank' => $authUserRank ? $authUserRank['rank'] : null,
+                
+        //     ];
+             
+        
+        //     return [
+        //         'date' => $group->first()->created_at->format('Y F'),
+        //         'authuser'=> $auth,
+        //         'top_list' => $rankedDrivers->take(10),
+        //     ];
+        // });
+        
+     
+        // return response()->json($trips);
+        
+        
+
+
+
+
+
+
+
+
+$startDate = Carbon::now()->startOfYear();
+$endDate = Carbon::now()->endOfYear();
+
+// Fetch trips with completed status within the date range
+$trips = Trip::where('status','completed')
+    ->whereBetween('created_at', [$startDate, $endDate])
+    ->orderBy('created_at')
+    ->get();
+
+$collection = collect($trips);
+
+// Group trips by month
+$trips = $collection->groupBy(function ($trip) {
+    return Carbon::parse($trip->created_at)->format('Y-m');
+})->map(function ($group) {
+    // Group by driver_id and get drivers with total trips
+    $drivers = $group->groupBy('driver_id')->map(function ($driverTrips, $driverId) {
+        $user = User::find($driverId);
+
+        // Check if the user exists and has the 'user' role
+        if ($user && $user->roles()->where('name', 'user')->exists()) {
+            return [
+                'driver_id' => $user->driver_id,
+                'name' => $user->name,
+                'total_trips' => $driverTrips->count(),
+                'profile_image' => optional($user->userImage)->profile_image,
+            ];
+        }
+        return null;
+    })->filter()->sortByDesc('total_trips')->values(); // Sort and re-index
+
+    // Add ranking to drivers
+    $rankedDrivers = $drivers->map(function ($driver, $index) {
+        $driver['rank'] = $index + 1; // Add rank based on index
+        return $driver;
+    }); // Take top 10
+
+    $authUser = Auth::user();
+    $authUserRank = $rankedDrivers->firstWhere('driver_id', $authUser->driver_id);
+
+
+    return [
+        'date' => $group->first()->created_at->format('Y F'),
+        'authuser' => $authUserRank,
+        'top_list' => $rankedDrivers->take(10),
+    ];
+});
+
+
+
+return response()->json($trips);
+
+        
+    }
     
 }
